code/cli/ai_ascii_art_live/requirements.txt
python-dotenv
openai
rich
aiofiles